Improvements to NetworkDriveMapper
The color of Drive changes on the MainPage when it is connected. Green for connected and red for disconnected.

Added the disconnect all drives button and moved the add a drive button to the ToolbarItems.

In the NetUseService, if the local device name is already taken, it will show this as a popup.