Improvements in GUI.
Fixed a bug with the progressbar and progress label not being adjusted properly.

If you click ''connect drive'' on a drive that is already connected . It will give you a popup that it is already connected and vise versa.

Added a fixed width and height for the app, code in AppShell.xaml.cs