GUI improvements and a bit of Code Refactoring.
Improved the GUI to show less data of the drive to keep the information safe.

Added form protection (Ex: The Letter must not be longer than a character or field shoudn't be empty)

Cover up the IP address and username as password box so the data cannot be copied or be seen.